Separate API from code
